Strip emails from arguments sent to analytics
We don't use email addresses in any path segments, query strings, or event
tracking on the site at the moment, but that doesn't mean that we don't
end up tracking them if they're added to incoming urls.  We try to detect
any potential email addresses in any arguments sent to our analytics
trackers and swap them out for the placeholder `[email]`.  We took
inspiration from similar work in govuk_admin_template (see:
https://github.com/alphagov/govuk_admin_template/pull/121) although
our regexp is different.  The admin version uses `\S@\S` but this is too
greedy for our purposes, as it would turn a path like
`/talk/to/government/embedded.email@example.com` into `[email]`, when we
want to keep theh rest of the path, just strip out the email address. We
instead use `[^\s=/?&]@[^\s=/?&]` which allows us to stop the regexp at
standard url path or query string boundaries.

We only use `stripPII` in the `trackPageview`, `trackEvent`, `trackShare`,
and `setDimension` methods and leave `addLinkedTrackerDomain` as it was.
`addLinkedTrackerDomain` sets config, rather than sends data to the
tracker so it's safe to leave as is (and it's not clear what might happen
if we interfered with the arguments to this method).  It doesn't look like
any code on github uses `sendToTrackers` directly, so we should be covered by
stripping PII only in those methods.