Strip postcodes out of data we send to trackers
Similar to the previous commit with emails we want to remove postcodes
from the data we send to analytics.  We re-use a (slightly modified)
version of the regexp from the [uk_postcodes gem][1] to match a postcode
in the arguments, strip it out, and replace it with `[postcode]`.  The
modifications are to only match a full postcode compared to the gem which
will match partials.  Partial postcodes aren't PII as they cover much
larger areas, so it shouldn't be a problem.

[1]: https://github.com/threedaymonk/uk_postcode/blob/master/lib/uk_postcode/geographic_postcode.rb#L13-L19